refactor: alterando Quantity de int para decimal

+++ src/PortfolioHub.Domain/Exceptions/ValueObjects/InsufficientQuantityException.cs
@@ -3,7 +3,7 @@ namespace PortfolioHub.Domain.Exceptions.ValueObjects;
 public class InsufficientQuantityException(string message) : BaseException(message)
 {
 
-    public static void ThrowIfInvalid(int value, int newValue)
+    public static void ThrowIfInvalid(decimal value, decimal newValue)
     {
         if (!(value >= newValue))
             ThrowIf(true, new InsufficientQuantityException("Não é possível subtrair!"));
